Connect to devices based on AD.
This patch adds code such that devices that may not have GATT service
objects created for them but the HR service UUID is listed in the
device's "uuids" field are also included in the drop down menu. Selecting
a device through the menu now explicitly opens a connection to the device.

The onDeviceAdded and onDeviceRemoved events are now observed as well.